README
Do you need a faster/simpler way of doing AI or ML without a GPU that can be run in real time. Then welcome to the Progressive Logic Unit Subroutine, or PLUSr for short.  The advantages of the PLUSr are it only needs integer mathematics to run, can be easily written in other simpler languages, be used on simpler processing systems or platforms, very fast that can learn in real time, processing and updating the AI or ML as the data comes in, great for robotics, uses all Boolean inputs, a Boolean output, and does not require a GPU. The disadvantages are all inputs needs to be binary (no continuous variable for input), because all inputs are binary it may be less accurate than other AI or ML systems that uses continuous variables for inputs.