Create PLUSr4
Do you need a faster/simpler way of doing AI or ML without a GPU that can be run in real time? Then welcome to the Progressive Logic Unit Subroutine, or PLUSr for short.  The advantages of the PLUSr are it- only needs integer mathematics to run, can be easily written in other simpler languages, can be used on simpler processing systems or platforms, can learn in real time very fast, processing and updating the AI or ML as the data comes in, great for robotics, uses all Boolean inputs, a Boolean output, and does not require a GPU. The disadvanages are- all inputs needs to be binary (no continuous variable for input), because all inputs are binary it mays be less accurate than other AI or ML systems that uses continuous variables for inputs. 

I hope you will be able to use and understand the PLUSr, even if you have only a passing knowlege of artifical intelligence, machine learning and neural networks.  I wrote the code in Python, so that it is easy to understand and use, but could been written in just about any language.

It is of substantial importance to state the benefits of PLUSr over other state-of-the-art AI programs (ex. Tensorflow, jax, and pytorch). The PLUSr completes its calculations in a fraction of the time it takes other programs AND it is able to make these calculations with minimal CPU usage, No GPU required AND STILL ABLE TO RUN HUNDREDS OF TIMES FASTER!